Bugfix: Look up a user referenced by an user-attribute-value only by its userId.
When a user-attribute-value is created, not only the userId is stored in the according field, but also the user-name is stored in value.
In the previous implementation the user was looked up by the stored user-name, instead of the userId.
If the name of the user had been changed, this resulted in an exception, because the user could not be found.

git-svn-id: http://scarab.tigris.org/svn/scarab@10483 3c7165bb-d2ef-0310-b724-bafd9a86081f