SCB2613: ResourceNotFoundException with urls containing the jsessionid:
Set the tomcat parameter useURIValidationHack to false.
This options prevents some security leaks in tomcat 4.0.x. Since we are using 4.1.x it should be ok to disable it.

Before the upgrade to tomcat v4.1.36 there was no problem with this option. 
After the upgrade the URIValidationHack causes that Request.decodedURI() contains the jsessionid-parameter. 
This URI is used, to extract the path for the velocity-template, which is used for the request. Because the extracted path is invalid, a ResourceNotFoundException is thrown.  

Disabling the URIValidationHack solves the problem.


git-svn-id: http://scarab.tigris.org/svn/scarab@10637 3c7165bb-d2ef-0310-b724-bafd9a86081f