cosmethic:
- Shortened an alert message:

old: This attribute is required to complete the information.
new: Required attribute must be set!

- moved alert messages from above the wrong field to below the field.

added a hook for javascript-support. Now each field contains an empty span
with a unique id:

<span id="${attributeName}:Hint" ></span>

This span has no impact on the GUI and currently it is just used from
extension methods (javascript add ons). As it is not functional by any means
i count it as cosmethic add on and added it here. (may be wrong)

git-svn-id: http://scarab.tigris.org/svn/scarab@10831 3c7165bb-d2ef-0310-b724-bafd9a86081f