SCB3015: CommonLinks.vm has overwritten the current data container by unconditionally changing the parameter queryId to the current default QueryId.
DetermineTargetValve.java sets the currentQueryId if needed. Here i have disregarded what happens when a query shall be EDITED instead of just called. The code now takes editing into account. In fact it now only modifies the event-parameter (to actually call a query), if queryId is available, but no event has been coded. In that case we have called a query using a short link.


git-svn-id: http://scarab.tigris.org/svn/scarab@10873 3c7165bb-d2ef-0310-b724-bafd9a86081f