SCB3036: There was a logical error in the last fix. Now the behaviour is as follows:
if a user has got a PRIVATE Query set as home page for this module, it is taken.
otherwise if there is a GLOBAL Query defined by any user (administrator) for this module, it is taken.

This is no good solution. Everybody should be able to select any query as his/her start page.
And if no query has been defeined by a user, he/she shall see a default start page as customized by an admin.
Probably this needs another field in the Query table which marks a query as the public default query. (to be checked)

git-svn-id: http://scarab.tigris.org/svn/scarab@10988 3c7165bb-d2ef-0310-b724-bafd9a86081f