SCB3036: Modified the Start_Page definition as follows:
If a user configures a personal query as start page, that query will be marked as such. If a public query exists for the same module, it will NOT be touched. Consequently now everybody who has npot configured a start page for the module, will get the public query. Everybody else will get his/her configured private query.

If a user (admin) configures a public query as start page, that query will be marked as such. Consequently now everybody who has not configured a start page for the module, will get the just configured public query. Everybody else will keep with his/her configured private query.

CAVEAT: public queries can not be customized for individual users. So whenever a user customizes a public query as start page, this customization will be taken as the new default settings for the module. It is not planned to change this behaviour as it is consistent with the idea of public/private queries.

git-svn-id: http://scarab.tigris.org/svn/scarab@10990 3c7165bb-d2ef-0310-b724-bafd9a86081f