Well, IE6 seems to have some problems with long line wrapping. It should work with all IE6 related fixes. But for some reason it breaks as soon as a white-space:pre-wrap is added to the css declaraion. The wrap itself works, but it does not free the wrapped space, hence the page width remains as if nowrap was choosen.
I fixed it by using a workaround for IE6 only: The comments section has now a width restriction (via css, can be customized in classic.css or custom.css)

While i was at it i added some divs for better styling. 

There was a distribution error in the radeox-jars. I also recompiled them and fixed a minor inconvenience directly related to this issue. As radeox seems to be non existing i again added the radeox sources (1.1-beta-2 now) as reference. and to be honest, i would rename it to radeox-1.1 (release). I do not see any more need for changes there...

git-svn-id: http://scarab.tigris.org/svn/scarab@11009 3c7165bb-d2ef-0310-b724-bafd9a86081f